smart-deal-app/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart
pymysql
cryptography


tinydb>=4.8.0
google-generativeai>=0.8.0
ollama>=0.6.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
pypdfium2>=4.24.0
Pillow>=10.0.0
pdf2image
numpy>=1.24.0
PyYAML>=6.0

